requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Vector Store local
chromadb>=0.5.0
langchain-chroma>=0.2.0

# Procesamiento de documentos
pymupdf>=1.24.0
python-docx>=1.1.0

# OCR para PDFs escaneados
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# Utilidades
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
python-telegram-bot>=21.0